+++ src/components/Map.tsx
@@ -1,11 +1,10 @@
-import { Feature, Geometry, MultiPoint } from "geojson";
+import { Feature, Geometry, Point } from "geojson";
 import { MapContainer, Marker, Popup, GeoJSON, TileLayer, CircleMarker as ReactCircleMarker } from 'react-leaflet'
 import { PathOptions, StyleFunction, LatLngExpression, CircleMarker, LeafletMouseEvent } from 'leaflet'
 import { useAppSelector, useAppDispatch } from "../app/hooks";
 import { TRACK_TYPE, ZONE_TYPE } from "../constants";
 import Track from "./Track";
 import Zone from "./Zone";
-import * as turf from "turf";
 import { useCallback } from "react";
 
 interface CustomPathOptions extends PathOptions {
@@ -16,39 +15,10 @@ const isVisible = (feature: Feature): boolean => {
   return feature.properties?.visible
 }
 
-const isTemporal = (feature: Feature): boolean => {
-  return feature.properties?.times
-}
-
-const timeVal = (timeStr: string): number => {
-  return new Date(timeStr).getTime()
-}
-
-const calcInterpLocation = (poly: MultiPoint, times: any, current: number, index: number): LatLngExpression => {
-  const coords = poly.coordinates
-  const isFirst = index === 0
-  const beforeIndex = isFirst ? 0 : index - 1
-  const afterIndex = isFirst ? 0 : index
-  const beforeCoords = coords[beforeIndex]
-  const afterCoords = coords[afterIndex]
-  const before = turf.point(beforeCoords)
-  const after = turf.point(afterCoords)
-  const turfPath = turf.lineString([beforeCoords, afterCoords])
-  const len = turf.distance(before, after)
-  const beforeTime = timeVal(times[beforeIndex])
-  const afterTime = timeVal(times[afterIndex])
-  const timeDelta = afterTime - beforeTime
-  const proportion = (current - beforeTime) / timeDelta
-  const lenProp = len * proportion
-  const interpolated = isNaN(lenProp) ? before : turf.along(turfPath, lenProp)
-  const markerLoc = interpolated.geometry.coordinates.reverse() as LatLngExpression
-  return markerLoc
-}
-
-
 const Map: React.FC = () => {
   const features = useAppSelector(state => state.featureCollection.features)
   const selectedFeaturesId = useAppSelector(state => state.selected.selected)
+  const currentLocations = useAppSelector(state => state.currentLocation.locations)
   const {current} = useAppSelector(state => state.time)
   const dispatch = useAppDispatch();
 
@@ -66,20 +36,6 @@ const Map: React.FC = () => {
     res.weight = 3
     return res;
   };
-
-  const InterpolatedLocationMarker = (feature: Feature, ctr: number, current: number): React.ReactElement => {
-    if (feature.properties?.times) {
-      const times = feature.properties.times
-      const index = times.findIndex((time: string) => new Date(time).getTime() >= current)
-      if (index >= 0) {
-        const poly = feature.geometry as MultiPoint
-        const markerLoc = calcInterpLocation(poly, times, current, index)
-        const key = `marker-${ctr}-${index}`
-        return <ReactCircleMarker key={key} radius={5} fillColor="#fff" color={feature.properties?.color || '#f9f'} center={markerLoc}/>
-      }
-    }
-    return <></>
-  }
 
   const onClickHandler = useCallback((id: string, modifier: boolean): void => {
     if (modifier) {
@@ -131,7 +87,12 @@ const Map: React.FC = () => {
           A pretty CSS3 popup. <br /> Easily customizable.
           </Popup>
         </Marker>
-        { current && features.filter(isTemporal).map((feature, ctr) => InterpolatedLocationMarker(feature, ctr, current)) }
+        {/* { current && currentLocations.map((feature: Feature<Point>) => {
+          const coords = feature.geometry.coordinates.slice().reverse() as LatLngExpression
+          const color = feature.properties?.color || '#f9f'
+          return <ReactCircleMarker key={feature.id as string + '-' + current } radius={5} 
+            fillColor={color} color={color} center={coords}/>}
+            )} */}
       </MapContainer>
     </>
   );
